Separate notebook-http and jupyter-websocket tests
* Rename test classes to something more reasonable in both cases
* Update Makefile TEST parameter
* Fix a masked test along the way

(c) Copyright IBM Corp. 2016
